Fix spelling of its vs it's in source and documentation.
To disabiguate the two homophones, the possessive form
is spelled "its", without an appostrophe. The contraction
"it's" retains the appostrophe. A good way to check which
to write is to try replacing "it's" with the expansion "it is".
If that doesn't make sense, it's the possessive, and should
instead be "its".